requirements.txt
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
bitsandbytes==0.48.2
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datasets==4.3.0
dill==0.4.0
distlib==0.4.0
einops==0.8.1
executing==2.2.1
facexlib==0.3.0
filelock==3.20.0
filterpy==1.4.5
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
GPUtil==1.4.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
icecream==2.1.8
identify==2.6.15
idna==3.11
imageio==2.37.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
lpips==0.1.4
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
nodeenv==1.9.1
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-clip==1.0.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
Pygments==2.19.2
pyiqa==0.1.14.1
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
ruff==0.14.2
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
spikezoo==0.2.3.7
spikingjelly==0.0.0.0.14
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.21
tokenizers==0.15.2
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.37.2
triton==3.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
